requirements.txt
Django==4.0.3
djangorestframework==3.13.1
environs==9.5.0
psycopg2==2.9.3
dj-database-url==0.5.0
phonenumbers==8.12.45
django-phonenumber-field==6.1.0
